re-structure file to isolate proxy awareness features from regular web client
